requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
airtrain==0.1.61
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
boto3==1.37.23
botocore==1.37.23
CacheControl==0.14.2
cachetools==5.5.2
cerebras_cloud_sdk==1.26.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
configuraptor==1.27.2
cryptography==44.0.2
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fake-useragent==2.1.0
filelock==3.18.0
firebase-admin==6.7.0
fireworks-ai==0.15.12
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.8.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.12.0
griffe==1.7.1
groq==0.20.0
grpcio==1.71.0
grpcio-status==1.71.0
gspread==6.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
loguru==0.7.3
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
modelcontextprotocol==0.1.0
monotonic==1.6
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
neo4j==5.28.1
notion-client==2.3.0
numpy==2.2.4
oauthlib==3.2.2
openai==1.69.0
openai-agents @ git+https://github.com/openai/openai-agents-python.git@9c53abe8c15ab2cf1c5591c1db1f61b52a1b24dc
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
plumbum==1.9.0
postgrest==1.0.1
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyairtable==3.0.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
requests==2.32.3
requests-oauthlib==2.0.0
result==0.17.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tabulate==0.9.0
threadful==0.5.0
together==1.5.4
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.11
uvicorn==0.34.0
uvx==1.0.3
wcwidth==0.2.13
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
